Tarea 3/notification-auto-sender/requirements.txt
attrs==25.3.0
blinker==1.9.0
click==8.1.8
flasgger==0.9.7.1
Flask==3.1.1
gunicorn==23.0.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mistune==3.1.3
packaging==25.0
pika==1.3.2
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.0
six==1.17.0
typing_extensions==4.13.2
Werkzeug==3.1.3
zipp==3.21.0
